requirements.txt
beautifulsoup4==4.9.3
Pillow==8.2.0
requests==2.25.1
selenium==3.141.0